requirements.txt
# Core Streamlit and UI
streamlit==1.28.1
plotly==5.17.0
pandas==2.1.3

# Azure SDK (Core)
azure-identity==1.15.0
azure-mgmt-containerservice==30.0.0
azure-mgmt-resource==23.0.1
azure-mgmt-compute==30.4.0
azure-keyvault-secrets==4.7.0

# LLM Providers
openai==1.3.7
google-generativeai==0.3.2
httpx==0.25.2

# Database and Vector Embeddings
chromadb==0.4.15
sentence-transformers==2.2.2

# Kubernetes
kubernetes==28.1.0
pyyaml==6.0.1

# Development and Utilities
python-dotenv==1.0.0
requests==2.31.0
rich==13.7.0
structlog==23.2.0
jinja2==3.1.2
jsonschema==4.20.0